requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   pytest
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
clip @ git+https://github.com/openai/CLIP.git
    # via -r requirements.in
fastapi==0.95.0
    # via -r requirements.in
filelock==3.10.7
    # via torch
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
ftfy==6.1.1
    # via clip
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via torch
markupsafe==2.1.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.0
    # via torch
numpy==1.24.2
    # via torchvision
openai==0.27.3
    # via -r requirements.in
packaging==23.0
    # via pytest
pillow==9.5.0
    # via torchvision
pluggy==1.0.0
    # via pytest
pydantic==1.10.7
    # via
    #   -r requirements.in
    #   fastapi
pytest==7.2.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
regex==2023.3.23
    # via clip
requests==2.28.2
    # via
    #   openai
    #   torchvision
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
sympy==1.11.1
    # via torch
torch==2.0.0
    # via
    #   -r requirements.in
    #   clip
    #   torchvision
torchvision==0.15.1
    # via
    #   -r requirements.in
    #   clip
tqdm==4.65.0
    # via
    #   clip
    #   openai
typing-extensions==4.5.0
    # via
    #   pydantic
    #   torch
urllib3==1.26.15
    # via requests
uvicorn==0.21.1
    # via -r requirements.in
wcwidth==0.2.6
    # via ftfy
yarl==1.8.2
    # via aiohttp
